requirements.txt
opencv-python~=4.10.0.84
numpy~=2.0.1
torchvision~=0.19.1
pillow~=10.4.0